% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_template.R
\name{create_new_proposal}
\alias{create_new_proposal}
\title{Create a new proposal}
\usage{
create_new_proposal(dir)
}
\arguments{
\item{dir}{Directory for the proposal}
}
\description{
Init a new directory for a proposal.
}
\examples{
/dontrun{
create_new_proposal("potential-client")
}

}
