% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse.R
\name{download_cleanshot_video}
\alias{download_cleanshot_video}
\title{Download video from Cleanshot}
\usage{
download_cleanshot_video(cleanshot_url, file_name)
}
\arguments{
\item{file_name}{}
}
\value{
Location of where video is downloaded
}
\description{
Download video from Cleanshot
}
