% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obtn_data.R
\name{download_file}
\alias{download_file}
\title{Download a file from a GitHub repository}
\usage{
download_file(file_path, dest_path, verbose = FALSE)
}
\arguments{
\item{file_path}{A character string specifying the path to the file in the GitHub repository.}

\item{dest_path}{A character string specifying the destination directory where the file will be saved.}

\item{verbose}{A logical value indicating whether to print messages during the download process. Default is \code{FALSE}.}
}
\value{
A character string representing the path to the downloaded file, or \code{NULL} if the download fails.
}
\description{
This function downloads a file from a specified GitHub repository and saves it to a local directory.
}
