% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse.R
\name{download_latest_zoom_video}
\alias{download_latest_zoom_video}
\title{Download Latest Zoom Recording}
\usage{
download_latest_zoom_video()
}
\value{
Returns location of video
}
\description{
Download Latest Zoom Recording
}
