% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus.R
\name{get_acs_race_ethnicity}
\alias{get_acs_race_ethnicity}
\title{Get race/ethnicity data from ACS}
\usage{
get_acs_race_ethnicity(...)
}
\arguments{
\item{...}{}
}
\value{
data frame with race/ethnicity data
}
\description{
Get race/ethnicity data from ACS
}
