% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_data.R
\name{get_blog_posts}
\alias{get_blog_posts}
\title{Get Blog Posts}
\usage{
get_blog_posts(statamic_token = Sys.getenv("STATAMIC_TOKEN"))
}
\arguments{
\item{statamic_token}{defaults to \code{Sys.getenv("STATAMIC_TOKEN")}}
}
\value{
data frame with all blog posts
}
\description{
Get Blog Posts
}
