% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin3.R
\name{get_course_ids}
\alias{get_course_ids}
\title{Get course ids}
\usage{
get_course_ids(course_title)
}
\arguments{
\item{course_title}{Title of course (or string to find multiple courses such as "R in 3 Months")}
}
\value{
ids of courses
}
\description{
Get course ids
}
