% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin3.R
\name{get_course_participants}
\alias{get_course_participants}
\title{Get all participants in a course or courses}
\usage{
get_course_participants(course_title)
}
\arguments{
\item{course_title}{Title of course (or string to find multiple courses such as "R in 3 Months")}
}
\description{
Get all participants in a course or courses
}
