% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obtn_data.R
\name{get_files}
\alias{get_files}
\title{Get a list of files matching a pattern from a GitHub repository}
\usage{
get_files(pattern, dir = "_targets/objects/")
}
\arguments{
\item{pattern}{A character string specifying the pattern to match file names.}

\item{dir}{A character string specifying the directory in the GitHub repository to search. Default is \code{"_targets/objects/"}.}
}
\value{
A character vector of file paths that match the pattern, or an error if the directory listing fails.
}
\description{
This function retrieves a list of files from a specified GitHub repository that match a given pattern.
}
