% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_data.R
\name{get_orders}
\alias{get_orders}
\title{Get all orders}
\usage{
get_orders(statamic_token = Sys.getenv("STATAMIC_TOKEN"))
}
\arguments{
\item{statamic_token}{defaults to \code{Sys.getenv("STATAMIC_TOKEN")}}
}
\value{
data frame with all orders
}
\description{
Get all orders
}
