% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social-media.R
\name{pdf_to_gif}
\alias{pdf_to_gif}
\title{Convert PDF to Animated GIF}
\usage{
pdf_to_gif(
  pdf_location,
  gif_file_path = here::here(),
  gif_file_name,
  density = 50
)
}
\arguments{
\item{pdf_location}{Location of PDF file}

\item{gif_file_path}{Starts at here::here()}

\item{gif_file_name}{You must include .gif at end}

\item{density}{Affects file size of GIF that is outputted}
}
\value{
GIF
}
\description{
Convert PDF to Animated GIF
}
