% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin3.R
\name{post_course_participants}
\alias{post_course_participants}
\title{Post participants in a course or courses to a Google Sheet}
\usage{
post_course_participants(
  course_title,
  google_sheets_url,
  sheet_name,
  view_sheet = TRUE
)
}
\arguments{
\item{course_title}{Title of course}

\item{google_sheets_url}{URL of Google Sheet to post to}

\item{sheet_name}{Name of sheet to post to}

\item{view_sheet}{Whether to open sheet in browser (TRUE by default)}
}
\value{
Posts participant info to Google Sheet
}
\description{
Post participants in a course or courses to a Google Sheet
}
