% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin3.R
\name{post_participant_info}
\alias{post_participant_info}
\title{Title}
\usage{
post_participant_info(
  course_title,
  google_sheets_url,
  sheet_name,
  view_sheet = TRUE
)
}
\arguments{
\item{course_title}{Title of course (or string to find multiple courses such as "R in 3 Months")}

\item{google_sheets_url}{URL of Google Sheet to post to}

\item{sheet_name}{Name of sheet to post to}

\item{view_sheet}{Whether to open sheet in browser (TRUE by default)}
}
\description{
Title
}
