% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{rfortherestofus_table}
\alias{rfortherestofus_table}
\title{R for the Rest of Us GT table}
\usage{
rfortherestofus_table(data, base_font = "Inter", base_font_size = 12, ...)
}
\arguments{
\item{data}{A dataframe or tibble.}

\item{base_font}{Default font to use (e.g., "Inter").}

\item{base_font_size}{Base font size, default to 12 (px).}

\item{...}{Additional arguments passed to \code{\link[gt:tab_options]{gt::tab_options()}}}
}
\value{
A gt table
}
\description{
A GT table theme with R for the Rest of Us branding.
}
\examples{
\dontrun{
data.frame(
  Item = c("R in 3 months with Group Discount"),
  Participants = c(100),
  CostPerParticipant = c(100),
  Total = c(100)
) |>
  rename(`Cost Per Participant` = "CostPerParticipant") |>
  rfor_table() |>
  fmt_currency(columns = c("Cost Per Participant", "Total"))
}

}
