% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_rru}
\alias{theme_rru}
\title{RRU ggplot2 theme}
\usage{
theme_rru(
  show_grid_lines = TRUE,
  show_legend = TRUE,
  base_family = "Inter",
  title_family = "Inter",
  void = FALSE
)
}
\arguments{
\item{show_grid_lines}{Boolean to indicate whether to have grid_lines (TRUE by default)}

\item{show_legend}{Whether or not to show the legend (TRUE by default)}

\item{base_family}{Base font family (Inter by default). If you want to use a custom font, you need to register it first.}

\item{title_family}{Title font family (Inter by default). If you want to use a custom font, you need to register it first.}

\item{void}{logical, use \code{TRUE} if you want the plot to use theme_void() as its starting point (i.e. when you don't want any axes)}
}
\value{
Ggplot2 theme
}
\description{
Remember to import fonts before using this, for the fonts to be used: rruthemes::import_fonts()!
}
