% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_md.R
\name{use_agents_md}
\alias{use_agents_md}
\title{Create or overwrite an AGENTS.md configuration file}
\usage{
use_agents_md(sections, overwrite = FALSE)
}
\arguments{
\item{sections}{Character vector with sections to add in the \code{AGENTS.md}. Possible
values are "R", "Quarto", "Typst".}

\item{overwrite}{Logical. Whether to overwrite an existing file. Default is \code{FALSE}.}
}
\description{
Generates an \code{AGENTS.md} file with common code practices
(R, Quarto) for LLMs. By default, writes a new file unless \code{overwrite = TRUE}
is specified.
You basically create the file with \code{rfortherestofus::use_agents_md()} and then
tell your favorite AI to read it. You can also copy paste the content to use in
other interface.
}
\examples{
\dontrun{
rfortherestofus::use_agents_md("R")              # just R rules
rfortherestofus::use_agents_md("Quarto")         # just Quarto rules
rfortherestofus::use_agents_md("Typst")          # just Typst rules
rfortherestofus::use_agents_md(c("R", "Quarto")) # R and Quarto rules
}

}
