% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{use_air_toml}
\alias{use_air_toml}
\title{Create or overwrite an air.toml configuration file}
\usage{
use_air_toml(
  line_width = 80,
  indent_width = 2,
  indent_style = "space",
  line_ending = "auto",
  persistent_line_breaks = "true",
  exclude = "[]",
  default_exclude = "true",
  skip = "[]",
  file_name = "air.toml",
  overwrite = FALSE
)
}
\arguments{
\item{line_width}{Integer. Maximum line width. Default is \code{80}.}

\item{indent_width}{Integer or character. Indentation width. If character, it
will be wrapped in quotes. Default is \code{2}.}

\item{indent_style}{Character. Indentation style, either \code{"space"} or \code{"tab"}.
Default is \code{"space"}.}

\item{line_ending}{Character. Line ending style. Can be \code{"auto"}, \code{"lf"},
\code{"crlf"}, or \code{"cr"}. Default is \code{"auto"}.}

\item{persistent_line_breaks}{Character (\code{"true"} or \code{"false"}). Whether to
preserve existing line breaks. Default is \code{"true"}.}

\item{exclude}{Character. TOML array of paths/patterns to exclude. Default is \code{"[]"}.}

\item{default_exclude}{Character (\code{"true"} or \code{"false"}). Whether to use
default exclusion rules. Default is \code{"true"}.}

\item{skip}{Character. TOML array of paths/patterns to skip. Default is \code{"[]"}.}

\item{file_name}{Character. Name of the configuration file. Must be
\code{"air.toml"} or \code{".air.toml"}. Default is \code{"air.toml"}.}

\item{overwrite}{Logical. Whether to overwrite an existing file. Default is \code{FALSE}.}
}
\description{
Generates an \code{air.toml} or \code{.air.toml} configuration file with formatting
options for the \code{air} code formatter. By default, writes a new file unless
\code{overwrite = TRUE} is specified.
}
\examples{
\dontrun{
use_air_toml()

use_air_toml(overwrite=TRUE)
}

}
