% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus.R
\name{view_acs_variables}
\alias{view_acs_variables}
\title{View all ACS variables}
\usage{
view_acs_variables(year = 2023)
}
\arguments{
\item{year}{}
}
\value{
data frame with all ACS variables
}
\description{
View all ACS variables
}
