% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rin3.R
\name{get_course_participants}
\alias{get_course_participants}
\title{Get all participants in a course}
\usage{
get_course_participants(course_title)
}
\arguments{
\item{course_title}{Title of course}
}
\value{
All participants in course
}
\description{
Get all participants in a course
}
