% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_course_repo.R
\name{init_course_repo}
\alias{init_course_repo}
\title{Initialize a new course repository
You should set up the course_repo before in rfortherestofus organization}
\usage{
init_course_repo(repo_name, instructor)
}
\arguments{
\item{repo_name}{Name of the new repo}

\item{instructor}{Github name of the instructor}
}
\value{
Issues and milestones written on Github
}
\description{
Initialize a new course repository
You should set up the course_repo before in rfortherestofus organization
}
\examples{
\dontrun{
init_course_repo(repo_name = "course_test", instructor = "tvroylandt")
}
}
