% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse.R
\name{upload_muse_video}
\alias{upload_muse_video}
\title{Upload video to muse.ai}
\usage{
upload_muse_video(local_video)
}
\arguments{
\item{local_video}{Location of local video}
}
\value{

}
\description{
Upload video to muse.ai
}
