% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus.R
\name{view_census_variables}
\alias{view_census_variables}
\title{View all Census variables}
\usage{
view_census_variables(year = 2010)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
View all Census variables
}
